Extract out a base type for the file context builder, so I can make another one.
