+++ src/Core/Gibberish/AST/_1_Bare/Builders/FileParseAsRawStatementSequenceBuilder.cs
@@ -1,53 +1,37 @@
 using System;
-using System.Collections.Generic;
-using JetBrains.Annotations;
 
 namespace Gibberish.AST._1_Bare.Builders
 {
-	public class FileParseAsRawStatementSequenceBuilder : AstBuilder<LanguageConstruct>
+	public class FileParseAsRawStatementSequenceBuilder : FileParseBuilderBase
 	{
 		public FileParseAsRawStatementSequenceBuilder(Action<FileParseAsRawStatementSequenceBuilder> content)
 		{
 			content(this);
 		}
 
-		[NotNull]
-		public BlankLineBuilder BlankLine(int indentationDepth)
+		public override BlankLineBuilder BlankLine(int indentationDepth)
 		{
 			return _Remember(new BlankLineBuilder(PossiblySpecified<int>.WithValue(indentationDepth)));
 		}
 
-		public StatementBuilder Statement(string content)
+		public override StatementBuilder Statement(string content)
 		{
 			return _Remember(new StatementBuilder(content, PossiblySpecified<int>.WithValue(0)));
 		}
 
-		public RawBlockBuilder Block(string prelude, Action<RawBlockBuilder.PreludeBuilder> preludeOptions)
+		public override RawBlockBuilder Block(string prelude, Action<RawBlockBuilder.PreludeBuilder> preludeOptions)
 		{
 			return _Remember(new RawBlockBuilder(prelude, preludeOptions, 0));
 		}
 
-		public RawBlockBuilder Block(string prelude)
+		public override RawBlockBuilder Block(string prelude)
 		{
 			return Block(prelude, x => { });
 		}
 
-		public CommentBuilder CommentDefinition(int commentId, string content)
+		public override CommentBuilder CommentDefinition(int commentId, string content)
 		{
 			return _Remember(new CommentBuilder(commentId, content));
 		}
-
-		public override void BuildInto(List<LanguageConstruct> destination)
-		{
-			foreach (var builder in _contents) { builder.BuildInto(destination); }
-		}
-
-		private TBuilder _Remember<TBuilder>(TBuilder line) where TBuilder : AstBuilder<LanguageConstruct>
-		{
-			_contents.Add(line);
-			return line;
-		}
-
-		[NotNull] private readonly List<AstBuilder<LanguageConstruct>> _contents = new List<AstBuilder<LanguageConstruct>>();
 	}
 }
