+++ src/Core/Gibberish/Gibberish.csproj
@@ -70,6 +70,7 @@
     <Compile Include="AST\_1_Bare\Builders\BlankLineBuilder.cs" />
     <Compile Include="AST\_1_Bare\Builders\BlockBuilder.cs" />
     <Compile Include="AST\_1_Bare\Builders\BlockBuilderBase.cs" />
+    <Compile Include="AST\_1_Bare\Builders\FileParseBuilderBase.cs" />
     <Compile Include="AST\_1_Bare\Builders\RawBlockBuilder.cs" />
     <Compile Include="AST\_1_Bare\Builders\CommentBuilder.cs" />
     <Compile Include="AST\_1_Bare\CommentDefinition.cs" />
